requirements.txt
fastapi
uvicorn[standard]

qdrant-client

sentence-transformers
torch

langchain
langchain-core
langchain-community
langchain-mistralai
langchain-experimental
langchain-huggingface

transformers
huggingface-hub

python-dotenv
pydantic

numpy
scikit-learn

streamlit
requests
httpx
tqdm
markdown
psycopg2-binary
redis